Extracting AnswerableField to its own component and building StringField as subclass
- No tests yet for AnswerableField, they will mostly absorb unit tests for string field
