+++ app/templates/components/questionnaire/answerable-field.hbs
@@ -1,0 +1,1 @@
+{{yield}}
